Address review comments
Rename modin/_compat/io -> modin/_compat/core
Rename modin/_compat/pandas -> modin/_compat/pandas_api
