+++ modin/_compat/core/pd_common.py
@@ -16,8 +16,12 @@
 from modin._compat import PandasCompatVersion
 
 if PandasCompatVersion.CURRENT == PandasCompatVersion.PY36:
-    from .py36.pd_common import get_handle
-elif PandasCompatVersion.CURRENT == PandasCompatVersion.LATEST:
-    from .latest.pd_common import get_handle
+    from .py36.pd_common import get_handle, pd_pivot_table
+    from pandas.core.aggregation import reconstruct_func
 
-__all__ = ["get_handle"]
+
+elif PandasCompatVersion.CURRENT == PandasCompatVersion.LATEST:
+    from .latest.pd_common import get_handle, pd_pivot_table
+    from pandas.core.apply import reconstruct_func
+
+__all__ = ["get_handle", "reconstruct_func", "pd_pivot_table"]
