+++ modin/_compat/pandas_api/abc/series.py
@@ -11,12 +11,11 @@
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
 
-from .base import Python36CompatibleBasePandasDataset
-from .dataframe import Python36CompatibleDataFrame
-from .series import Python36CompatibilitySeries
+"""Module for interface for class compatibility layer for Series."""
 
-__all__ = [
-    "Python36CompatibleBasePandasDataset",
-    "Python36CompatibleDataFrame",
-    "Python36CompatibilitySeries",
-]
+from modin.logging import ClassLogger
+from .base import BaseCompatibilityBasePandasDataset
+
+
+class BaseCompatibilitySeries(BaseCompatibilityBasePandasDataset, ClassLogger):
+    """Interface for class compatibility layer for Series."""
