+++ modin/pandas/__init__.py
@@ -14,13 +14,28 @@
 import pandas
 import warnings
 
-__pandas_version__ = "1.4.3"
+from modin._compat import PandasCompatVersion
 
-if pandas.__version__ != __pandas_version__:
-    warnings.warn(
-        f"The pandas version installed {pandas.__version__} does not match the supported pandas version in"
-        + f" Modin {__pandas_version__}. This may cause undesired side effects!"
-    )
+if PandasCompatVersion.CURRENT == PandasCompatVersion.PY36:
+    __pandas_version__ = "1.1.5"
+
+    if pandas.__version__ != __pandas_version__:
+        warnings.warn(
+            f"The pandas version installed {pandas.__version__} does not match the supported pandas version in"
+            + f" Modin {__pandas_version__} compatibility mode. This may cause undesired side effects!"
+        )
+    else:
+        warnings.warn(
+            f"Starting Modin in compatibility mode to support legacy pandas version {__pandas_version__}"
+        )
+elif PandasCompatVersion.CURRENT == PandasCompatVersion.LATEST:
+    __pandas_version__ = "1.4.3"
+
+    if pandas.__version__ != __pandas_version__:
+        warnings.warn(
+            f"The pandas version installed {pandas.__version__} does not match the supported pandas version in"
+            + f" Modin {__pandas_version__}. This may cause undesired side effects!"
+        )
 
 with warnings.catch_warnings():
     warnings.simplefilter("ignore")
@@ -225,11 +240,11 @@ from .general import (
     wide_to_long,
 )
 
-from modin._compat.pandas.namespace import pivot_table
+from modin._compat.pandas_api.namespace import pivot_table
 from modin._compat import PandasCompatVersion
 
 if PandasCompatVersion.CURRENT != PandasCompatVersion.PY36:
-    from modin._compat.pandas.namespace import Flags, Float32Dtype, Float64Dtype
+    from modin._compat.pandas_api.namespace import Flags, Float32Dtype, Float64Dtype
 del PandasCompatVersion
 
 from .plotting import Plotting as plotting
