+++ modin/pandas/base.py
@@ -49,7 +49,7 @@ from modin import pandas as pd
 from modin.pandas.utils import is_scalar
 from modin.config import IsExperimental
 from modin.logging import disable_logging
-from modin._compat.pandas.classes import BasePandasDatasetCompat
+from modin._compat.pandas_api.classes import BasePandasDatasetCompat
 
 # Similar to pandas, sentinel value to use as kwarg in place of None when None has
 # special meaning and needs to be distinguished from a user explicitly passing None.
@@ -953,7 +953,7 @@ class BasePandasDataset(BasePandasDatasetCompat):
         return self.loc[indexer] if axis == 0 else self.loc[:, indexer]
 
     @_inherit_docstrings(
-        pandas.DataFrame.between_time, apilink="pandas.DataFrame,between_time"
+        pandas.DataFrame.between_time, apilink="pandas.DataFrame.between_time"
     )
     def _between_time(self, **kwargs):
         axis = self._get_axis_number(kwargs.pop("axis", None))
